Lab01

+++ src/main/java/inventory/controller/MainScreenController.java
@@ -138,7 +138,7 @@ public class MainScreenController implements Initializable,Controller {
         alert.setContentText("Are you sure you want to delete part " + part.getName() + " from parts?");
         Optional<ButtonType> result = alert.showAndWait();
 
-        if (result.get() == ButtonType.OK) {
+        if (result.isPresent() && result.get() == ButtonType.OK) {
             System.out.println("Part deleted.");
             service.deletePart(part);
         } else {
